Extract visibility culling into a Culler class
Move per-camera light visibility, mesh-instance culling (request/execute) and shadow-caster
culling out of the long Renderer into a new Culler (src/scene/renderer/culler.js). The renderer
owns a Culler instance (renderer.culler) and delegates to it; internal callers go through it.
Shared utilities (updateCameraFrustum, collectLights) and the cull stats stay on the renderer.
Pure structural refactor, no behavior change.
